Implemented Splitting and Encapsulated File IO
Now the program writes the adjusted (based on the MOST RECENT stock price) price of stocks to file. In addition, new methods have been created for file input output, so it is in the class instead of in the "main" function.

Also, I've elected to ignore E501 in PEP8 by extending the line count to 200 chars.